First working version
This is the first functional graphical front-end for nokyc.py. There is no graphical indication of the tor status, but as long as the file is started from the command line, the tor status will output to that command line. 

This project currently runs tor and nokyc.py in a container. The reason for this, is so that tor does not have to run system wide. Open to suggestions on how to better handle the tor dependency for nokyc.py